requirements.txt
pytest==7.4.0
requests==2.32.0
PyYAML==6.0.1
PyMySQL==1.1.0
allure-pytest==2.13.2
jsonschema==4.19.0
pytest-html==4.1.1
